refactor to remove productReview context
