+++ client/src/components/Ratings/Ratings.jsx
@@ -1,10 +1,8 @@
 import React from 'react';
 import styled from 'styled-components';
 import { useMeta } from '../../contexts/ReviewMeta';
-
 import RatingBreakdown from './RatingBreakdown';
 import ReviewList from './ReviewList';
-import { ReviewProvider } from '../../contexts/ProductReview';
 
 export default function Ratings() {
   const currentMeta = useMeta();
@@ -13,9 +11,7 @@ export default function Ratings() {
   return (
     <RatingSection>
       {ratingBreakdown}
-      <ReviewProvider>
-        <ReviewList />
-      </ReviewProvider>
+      <ReviewList />
     </RatingSection>
   );
 }
